@@ -24,7 +24,7 @@
  */
 package fr.opensagres.xdocreport.document.web;
 
-import javax.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletRequest;
 
 import fr.opensagres.xdocreport.converter.IURIResolver;
 
