@@ -27,7 +27,7 @@ package fr.opensagres.xdocreport.service.rest;
 import java.util.HashSet;
 import java.util.Set;
 
-import javax.ws.rs.core.Application;
+import jakarta.ws.rs.core.Application;
 
 public class XDocreportApplication
     extends Application
