@@ -24,10 +24,10 @@
  */
 package fr.opensagres.xdocreport.service.ws;
 
-import javax.servlet.ServletConfig;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
-import javax.xml.ws.Endpoint;
+import jakarta.servlet.ServletConfig;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.xml.ws.Endpoint;
 
 public class WSXDocReportServlet
     extends HttpServlet
