@@ -24,32 +24,6 @@
  */
 package fr.opensagres.xdocreport.document.odt.styling;
 
-import static org.junit.Assert.assertThat;
-
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.StringReader;
-import java.io.StringWriter;
-import java.io.Writer;
-
-import javax.xml.parsers.DocumentBuilder;
-import javax.xml.parsers.DocumentBuilderFactory;
-import javax.xml.transform.OutputKeys;
-import javax.xml.transform.Transformer;
-import javax.xml.transform.TransformerFactory;
-import javax.xml.transform.dom.DOMSource;
-import javax.xml.transform.stream.StreamResult;
-
-import org.junit.Assert;
-import org.junit.Test;
-import org.w3c.dom.Document;
-import org.xml.sax.InputSource;
-import org.xml.sax.XMLReader;
-import org.xml.sax.helpers.XMLReaderFactory;
-import org.xmlunit.matchers.CompareMatcher;
-
-import com.sun.org.apache.xml.internal.serializer.OutputPropertiesFactory;
-
 import fr.opensagres.xdocreport.document.odt.preprocessor.ODTBufferedDocumentContentHandler;
 import fr.opensagres.xdocreport.document.odt.preprocessor.ODTStyleContentHandler;
 import fr.opensagres.xdocreport.document.odt.textstyling.MockContext;
@@ -59,6 +33,28 @@ import fr.opensagres.xdocreport.document.preprocessor.sax.BufferedDocumentConten
 import fr.opensagres.xdocreport.document.textstyling.IDocumentHandler;
 import fr.opensagres.xdocreport.document.textstyling.ITextStylingTransformer;
 import fr.opensagres.xdocreport.document.textstyling.html.HTMLTextStylingTransformer;
+import org.junit.Assert;
+import org.junit.Test;
+import org.w3c.dom.Document;
+import org.xml.sax.InputSource;
+import org.xml.sax.XMLReader;
+import org.xml.sax.helpers.XMLReaderFactory;
+import org.xmlunit.matchers.CompareMatcher;
+
+import javax.xml.parsers.DocumentBuilder;
+import javax.xml.parsers.DocumentBuilderFactory;
+import javax.xml.transform.OutputKeys;
+import javax.xml.transform.Transformer;
+import javax.xml.transform.TransformerFactory;
+import javax.xml.transform.dom.DOMSource;
+import javax.xml.transform.stream.StreamResult;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.StringReader;
+import java.io.StringWriter;
+import java.io.Writer;
+
+import static org.junit.Assert.assertThat;
 
 /**
  * Check ODT Styling generation by comparing generated result against the expected XML
@@ -106,7 +102,7 @@ public class TestODTStyling
             Transformer serializer = tfactory.newTransformer();
 
             serializer.setOutputProperty( OutputKeys.INDENT, "yes" );
-            serializer.setOutputProperty( OutputPropertiesFactory.S_KEY_INDENT_AMOUNT, "2" );
+            serializer.setOutputProperty( "{http:// xml. apache. org/ xalan}indent-amount", "2" );
             serializer.setOutputProperty( OutputKeys.STANDALONE, "no" );
 
             Writer out = new StringWriter();
