@@ -80,7 +80,7 @@ public interface IImageRegistry
 
     /**
      * Returns the width of the image provider if not null and the given defaultWidth otherwise.
-     * 
+     *
      * @param imageProvider
      * @param defaultWidth
      * @return
@@ -90,13 +90,37 @@ public interface IImageRegistry
         throws IOException;
 
     /**
-     * Returns the height of the image provider if not null and the given defaultHeight otherwise.
+     * Returns the width of the image provider if not null and the given defaultWidth otherwise.
      * 
      * @param imageProvider
      * @param defaultWidth
+     * @param defaultHeight
+     * @return
+     * @throws IOException
+     */
+    String getWidth( ImageProviderInfo info, String defaultWidth, String defaultHeight )
+        throws IOException;
+
+    /**
+     * Returns the height of the image provider if not null and the given defaultHeight otherwise.
+     *
+     * @param imageProvider
+     * @param defaultHeight
      * @return
      * @throws IOException
      */
     String getHeight( ImageProviderInfo info, String defaultHeight )
+        throws IOException;
+
+    /**
+     * Returns the height of the image provider if not null and the given defaultHeight otherwise.
+     * 
+     * @param imageProvider
+     * @param defaultWidth
+     * @param defaultHeight
+     * @return
+     * @throws IOException
+     */
+    String getHeight( ImageProviderInfo info, String defaultWidth, String defaultHeight )
         throws IOException;
 }
